testing/web-platform/tests/workers/WorkerLocation_protocol.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerLocation URL decomposition IDL attribute: protocol </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = './support/WorkerLocation.js';
    var ExpectedResult = window.location.protocol;
            
    var t = async_test("Test Description: WorkerLocation protocol attribute returns the current scheme of the underlying URL.");
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data.protocol, ExpectedResult);
        t.done();
    });
</script>
</body>
</html>
